+++ examples/appliance_health_status.py
@@ -30,14 +30,9 @@ config = {
 # Try load config from a file (if there is a config file)
 config = try_load_from_file(config)
 oneview_client = OneViewClient(config)
-app_node_info = oneview_client.appliance_node_information
+app_health = oneview_client.appliance_health_status
 
-# Get node status information from appliance
-print("\nGet node status information from appliance:\n ")
-node_status = app_node_info.get_status()
-pprint(node_status.data)
-
-# Get node version information from appliance
-print("\nGet node version information from appliance\n")
-node_version = app_node_info.get_version()
-pprint(node_version.data)
+# Get health status information from appliance
+print("\nGet health status information from appliance:\n ")
+health_status = app_health.get_health_status()
+pprint(health_status.data)