+++ tests/unit/resources/settings/test_appliance_health_status.py
@@ -20,22 +20,17 @@ import unittest
 import mock
 
 from hpeOneView.connection import connection
-from hpeOneView.resources.settings.appliance_node_information import ApplianceNodeInformation
+from hpeOneView.resources.settings.appliance_health_status import ApplianceHealthStatus
 from hpeOneView.resources.resource import Resource
 
 
-class ApplianceNodeInformationTest(unittest.TestCase):
+class ApplianceHealthStatusTest(unittest.TestCase):
     def setUp(self):
         self.host = '127.0.0.1'
         self.connection = connection(self.host, 800)
-        self._node_information = ApplianceNodeInformation(self.connection)
+        self._health_status = ApplianceHealthStatus(self.connection)
 
     @mock.patch.object(Resource, 'get_by_uri')
-    def test_get_status_called_once(self, mock_get):
-        self._node_information.get_status()
-        mock_get.assert_called_once_with('/rest/appliance/nodeinfo/status')
-
-    @mock.patch.object(Resource, 'get_by_uri')
-    def test_get_version_called_once(self, mock_get):
-        self._node_information.get_version()
-        mock_get.assert_called_once_with('/rest/appliance/nodeinfo/version')
+    def test_get_health_status_called_once(self, mock_get):
+        self._health_status.get_health_status()
+        mock_get.assert_called_once_with('/rest/appliance/health-status')